requirements.txt
fastapi
uvicorn
pymongo
playwright
huggingface_hub
scikit-learn
requests
lxml
python-dotenv
redis
